move perf helper to a perf file
